Show off Rust's uniqueness
Foo and Bar are so last year. To show that we are a language to
count on in 2017, we need new fresh terms that remind us
of what language we're dealing with.

We should also do the same on all crates on crates.io. I suggest
setting up a http://arewefermentedyet.com/ website to track the purpose.